added information regarding required python version
This is a wonderful library is supposed to work with python 2.7.xx
I didn't know it initially and discovered a syntax unfamiliar to me in the print statements.

It would be helpful for python3 users like me to learn how to make use of legacy scripts. 

Adding this information to the documentation should help.